Add files via upload
The async needs to be fixed now, but this revision allows you to switch from a search page to a current playlist page. Need to fix the async before formatting the playlist page more. For some reason the app geometry is also messed up, I don't know why. This revision uses tk.raise() to alternate between the pages.